add fragments and api bundle to support different terminal versions
