+++ .mvn/maven.config
@@ -1,0 +1,1 @@
+-Dtycho.localArtifacts=ignore
