+++ base.target
@@ -3,7 +3,7 @@
 <target includeMode="feature" name="Copilot for Eclipse Target Definition">
 	<locations>
 		<location includeAllPlatforms="false" includeConfigurePhase="false" includeMode="planner" includeSource="true" type="InstallableUnit">
-			<repository location="https://download.eclipse.org/releases/2024-12/"/>
+			<repository location="https://download.eclipse.org/releases/2025-03/"/>
 			<unit id="org.eclipse.equinox.executable.feature.group" version="0.0.0"/>
 			<unit id="org.eclipse.platform.ide" version="0.0.0"/>
 			<unit id="org.eclipse.jdt.feature.group" version="0.0.0"/>
@@ -12,21 +12,17 @@
 			<unit id="org.apache.commons.lang3" version="0.0.0"/>
 			<unit id="org.eclipse.tm4e.feature.feature.group" version="0.0.0"/>
 			<unit id="org.eclipse.tm4e.language_pack.feature.feature.group" version="0.0.0"/>
-			<unit id="org.eclipse.tm.terminal.feature.feature.group" version="0.0.0"/>
-			<unit id="org.eclipse.tm.terminal.view.feature.feature.group" version="0.0.0"/>
-			<unit id="org.eclipse.tm.terminal.control.feature.feature.group" version="0.0.0"/>
-			<unit id="org.eclipse.tm.terminal.connector.local.feature.feature.group" version="0.0.0"/>
 			<unit id="org.eclipse.egit.feature.group" version="0.0.0"/>
 			<unit id="org.eclipse.jgit" version="0.0.0"/>
+		</location>
+		<location includeAllPlatforms="false" includeConfigurePhase="false" includeMode="planner" includeSource="true" type="InstallableUnit">
+			<repository location="https://download.eclipse.org/cbi/updates/license/"/>
+			<unit id="org.eclipse.license.feature.group" version="0.0.0"/>
 		</location>
 		<location includeAllPlatforms="false" includeConfigurePhase="false" includeMode="planner" includeSource="true" type="InstallableUnit">
 				<!-- Wildwebdeveloper version 1.3.2 is the latest version that support JavaSE-17 as the execution environment -->
 				<repository location="https://download.eclipse.org/wildwebdeveloper/releases/1.3.2/"/>
 				<unit id="org.eclipse.wildwebdeveloper.embedder.node.feature.feature.group" version="0.0.0"/>
-		</location>
-		<location includeAllPlatforms="false" includeConfigurePhase="false" includeMode="planner" includeSource="true" type="InstallableUnit">
-				<repository location="https://download.eclipse.org/cbi/updates/license/"/>
-				<unit id="org.eclipse.license.feature.group" version="0.0.0"/>
 		</location>
 		<location includeAllPlatforms="false" includeConfigurePhase="false" includeMode="planner" includeSource="true" type="InstallableUnit">
 			<repository location="https://download.eclipse.org/mylyn/updates/release/4.4.0"/>
@@ -55,36 +51,25 @@
 				</dependency>
 			</dependencies>
 		</location>
-		<location includeAllPlatforms="false" includeConfigurePhase="true" includeMode="planner"
-			includeSource="true" type="InstallableUnit">
-			<repository
-				location="https://download.eclipse.org/tools/orbit/downloads/drops/R20220302172233/repository/" />
-			<unit id="org.w3c.css.sac" version="1.3.1.v200903091627" />
-			<unit id="org.w3c.dom.svg" version="1.1.0.v201011041433" />
-			<unit id="org.w3c.dom.smil" version="1.0.1.v200903091627" />
-			<unit id="org.w3c.dom.events" version="3.0.0.draft20060413_v201105210656" />
+		<location includeAllPlatforms="false" includeConfigurePhase="true" includeMode="planner" includeSource="true" type="InstallableUnit">
+			<repository location="https://download.eclipse.org/tools/orbit/downloads/drops/R20220302172233/repository/"/>
+			<unit id="org.w3c.css.sac" version="1.3.1.v200903091627"/>
+			<unit id="org.w3c.dom.svg" version="1.1.0.v201011041433"/>
+			<unit id="org.w3c.dom.smil" version="1.0.1.v200903091627"/>
+			<unit id="org.w3c.dom.events" version="3.0.0.draft20060413_v201105210656"/>
 		</location>
 		<!-- E4 Tools including CSS Spy -->
 		<!-- E4 CSS Editor and Tools -->
-		<location includeAllPlatforms="false" includeConfigurePhase="false" includeMode="planner"
-			includeSource="true" type="InstallableUnit">
-			<repository
-				location="https://download.eclipse.org/e4/snapshots/org.eclipse.e4.tools/latest/" />
-			<unit id="org.eclipse.e4.tools.css.spy.feature.feature.group"
-				version="0.14.0.v20200923-1237" />
-			<unit id="org.eclipse.e4.tools.bundle.spy.feature.feature.group"
-				version="0.1.0.v20200923-1237" />
-			<unit id="org.eclipse.e4.tools.context.spy.feature.feature.group"
-				version="0.17.0.v20200923-1237" />
-			<unit id="org.eclipse.e4.tools.event.spy.feature.feature.group"
-				version="0.15.0.v20200923-1237" />
-			<unit id="org.eclipse.e4.tools.model.spy.feature.feature.group"
-				version="0.1.100.v20200923-1237" />
-			<unit id="org.eclipse.e4.tools.preference.spy.feature.feature.group"
-				version="0.1.0.v20200923-1237" />
-			<unit id="org.eclipse.e4.tools.spies.feature.feature.group"
-				version="0.18.0.v20200923-1237" />
+		<location includeAllPlatforms="false" includeConfigurePhase="false" includeMode="planner" includeSource="true" type="InstallableUnit">
+			<repository location="https://download.eclipse.org/e4/snapshots/org.eclipse.e4.tools/latest/"/>
+			<unit id="org.eclipse.e4.tools.css.spy.feature.feature.group" version="0.14.0.v20200923-1237"/>
+			<unit id="org.eclipse.e4.tools.bundle.spy.feature.feature.group" version="0.1.0.v20200923-1237"/>
+			<unit id="org.eclipse.e4.tools.context.spy.feature.feature.group" version="0.17.0.v20200923-1237"/>
+			<unit id="org.eclipse.e4.tools.event.spy.feature.feature.group" version="0.15.0.v20200923-1237"/>
+			<unit id="org.eclipse.e4.tools.model.spy.feature.feature.group" version="0.1.100.v20200923-1237"/>
+			<unit id="org.eclipse.e4.tools.preference.spy.feature.feature.group" version="0.1.0.v20200923-1237"/>
+			<unit id="org.eclipse.e4.tools.spies.feature.feature.group" version="0.18.0.v20200923-1237"/>
 		</location>
 	</locations>
-	<targetJRE path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-17"/>
+	<targetJRE path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-21"/>
 </target>
