+++ org.eclipse.copilot.api/META-INF/MANIFEST.MF
@@ -1,0 +1,12 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: org.eclipse.copilot.api
+Bundle-SymbolicName: org.eclipse.copilot.api
+Bundle-Version: 0.10.1.qualifier
+Export-Package: org.eclipse.copilot.api
+Automatic-Module-Name: org.eclipse.copilot.api
+Bundle-RequiredExecutionEnvironment: JavaSE-17
+Import-Package: org.osgi.framework;version="[1.10.0,2.0.0)"
+Require-Bundle: org.eclipse.ui;bundle-version="3.205.100",
+ org.apache.commons.lang3;bundle-version="3.14.0",
+ org.eclipse.core.runtime;bundle-version="[3.31.0,4.0.0)"
