+++ org.eclipse.copilot.api/META-INF/p2.inf
@@ -1,0 +1,11 @@
+requires.0.namespace=org.eclipse.equinox.p2.iu
+requires.0.name=org.eclipse.copilot.ui.terminal
+requires.0.greedy=true
+requires.0.optional=true
+requires.0.filter = (!(org.eclipse.copilot=true))
+
+requires.1.namespace=org.eclipse.equinox.p2.iu
+requires.1.name=org.eclipse.copilot.ui.terminal.legacy
+requires.1.greedy=true
+requires.1.optional=true
+requires.1.filter = (!(org.eclipse.copilot=true))
