+++ org.eclipse.copilot.feature/feature.xml
@@ -27,5 +27,8 @@
    <plugin
          id="org.eclipse.copilot.ui"
          version="0.0.0"/>
+   <plugin
+         id="org.eclipse.copilot.api"
+         version="0.0.0"/>
 
 </feature>
