+++ org.eclipse.copilot.ui.terminal/META-INF/services/org.eclipse.copilot.api.IRunInTerminalTool
@@ -1,0 +1,1 @@
+org.eclipse.copilot.ui.terminal.RunInTerminalTool
