+++ org.eclipse.copilot.ui/META-INF/MANIFEST.MF
@@ -22,12 +22,13 @@ Automatic-Module-Name: org.eclipse.copilot.ui
 Bundle-ActivationPolicy: lazy
 Import-Package: org.osgi.framework;version="[1.10.0,2.0.0)"
 Require-Bundle: org.eclipse.copilot.core;bundle-version="0.10.1",
+ org.eclipse.copilot.api;bundle-version="0.10.1",
  org.eclipse.ui.ide;bundle-version="3.22.100",
  org.eclipse.ui.workbench.texteditor;bundle-version="3.17.300",
  org.eclipse.ui;bundle-version="3.205.100",
  org.eclipse.ui.navigator;bundle-version="3.12.300",
  org.eclipse.jface.text;bundle-version="3.25.0",
- org.eclipse.core.runtime;bundle-version="3.31.0",
+ org.eclipse.core.runtime;bundle-version="[3.31.0,4.0.0)",
  org.eclipse.jdt.annotation;resolution:=optional,
  org.eclipse.core.resources;bundle-version="3.20.100",
  org.eclipse.lsp4e;bundle-version="0.18.6",
@@ -43,10 +44,6 @@ Require-Bundle: org.eclipse.copilot.core;bundle-version="0.10.1",
  org.eclipse.core.databinding.observable;bundle-version="1.13.200",
  org.eclipse.jface.databinding;bundle-version="1.15.200",
  com.google.gson;bundle-version="2.10.1",
- org.eclipse.tm.terminal.view.core,
- org.eclipse.tm.terminal.view.ui,
- org.eclipse.tm.terminal.control,
- org.eclipse.tm.terminal.connector.local,
  org.eclipse.compare;bundle-version="3.10.0",
  org.osgi.service.event;bundle-version="1.4.1",
  org.eclipse.e4.core.services;bundle-version="2.4.300",
