+++ target-legacy-terminal.target
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<?pde version="3.8"?>
+<target name="Copilot for Eclipse Target Definition - Terminal">
+	<locations>
+			<location includeAllPlatforms="false" includeConfigurePhase="false" includeMode="planner" includeSource="true" type="InstallableUnit">
+			<repository location="https://download.eclipse.org/releases/2024-12/"/>
+			<unit id="org.eclipse.tm.terminal.feature.feature.group" version="0.0.0"/>
+			<unit id="org.eclipse.tm.terminal.view.feature.feature.group" version="0.0.0"/>
+			<unit id="org.eclipse.tm.terminal.control.feature.feature.group" version="0.0.0"/>
+			<unit id="org.eclipse.tm.terminal.connector.local.feature.feature.group" version="0.0.0"/>
+		</location>
+		<location type="Target" uri="file:${project_loc:/eclipse-copilot}/base.target" />
+    </locations>
+</target>
